==> 01_Data_Collection_Again.ipynb <==
import praw
import pandas as pd
from tqdm import tqdm
import time

reddit = praw.Reddit(
    client_id="y25BJERt0d-YtQKuTk7gUQ",
    client_secret="fBb7prfE4DcIj2Cv4l7Dw73ziLZFyQ",
    user_agent="sl-nlp-project:v1 (by /u/SpareUse7114)", 
)
reddit.read_only

subreddit = reddit.subreddit("srilanka")
subreddit

keywords=[
  "A/l", "Abuse", "Admission", "Alevel", "Alumni", "Amantha", "Amaradeva", "Another", "Any", "Arjuna",
  "Assignment", "Bolt", "Boy", "Campus", "Cardinal", "Cars", "Certificate", "Charges", "College", "Coursework",
  "Curriculum", "Day", "Degree", "Department", "Dhanith", "Dhoni", "Dilshan", "Dimuth", "Diploma", "Drugs",
  "Education system", "Enrollment", "Every", "Exam", "Faculty", "Ganja","Marijuana", "Genz", "Get", "Girl", "Graduation",
  "Guy", "Hamilton", "Harini", "High", "Homework", "Ice", "Imran", "Institute", "Iraj", "Ishara",
  "Jacqueline", "Jerome", "Johnston", "Jothipala", "Kanjipaana", "Kindergarden", "Know", "Kohli", "Kosgoda", "Kush",
  "Last", "Learning", "Lecturer", "Life", "Madush", "Mahela", "Makandure", "Malcolm", "Malinga","Mom","Nvidia","Rysen"
  "Man", "Mathews", "Messi", "Montessori", "Month", "Murali", "Murder", "New", "Neymar", "Now","Dad",
  "O/l", "Offences", "Olevel", "Only", "Other", "Pastor", "People", "Phelps", "Preschool", "Primary",
  "Professor", "Project", "Ranjan", "Ranjith", "Really", "Rohit", "Ronaldo", "Sachin", "Sanga",
  "Scholarship", "Secondary", "See", "Sewwandi", "Sex", "Situationship", "Some", "Study", "Sunil perera", "Susanthika",
  "Take", "Test", "Tharaka", "Thesis", "Thing", "Time", "Today", "Tomorrow", "Trafficking", "Tuition",
  "Underworld", "Very", "Virat", "Way", "Week", "Woman", "Year", "Yesterday", "Yohani"
]


print(len(keywords), "keywords to search")

## Text Posts

limit_per_keyword = 1000  
posts_data = []

for keyword in keywords:
    for submission in tqdm(subreddit.search(keyword, limit=limit_per_keyword)):
        try:
            if submission.title.strip() == "" and submission.selftext.strip() == "":
                continue  
            
            posts_data.append({
                    "id": submission.id,
                    "post_title": submission.title,
                    "type": "text",
                    "body": submission.selftext,
                    "score": submission.score,
                    "url": submission.url,
                    "created_utc": submission.created_utc
            })
            
            time.sleep(0.5)  # respect Reddit API rate limit
        except Exception as e:
            print("Error:", e)
            continue

df_text_posts = pd.DataFrame(posts_data)
df_text_posts.head()

df_text_posts.info()

## Combine all datasets

df1 = pd.read_csv("raw data\srilanka_posts_text.csv")
df2 = pd.read_csv("raw data\srilanka_posts_non_text.csv")

combined_df = pd.concat([df1, df2, df_text_posts], ignore_index=True)

combined_df.info()

if not combined_df.empty:
    combined_df.to_csv("raw data\srilanka_all_text.csv", index=False, encoding="utf-8")
    print("Saved to raw data\srilanka_all_text.csv")
else:
    print("\nNo data.")
